+++ src/migrations/Install.php
@@ -10,39 +10,18 @@
 
 namespace venveo\oauthclient\migrations;
 
-use Craft;
 use craft\db\Migration;
 
-/**
- * @author    Venveo
- * @package   Oauth20Client
- * @since     1.0.0
- */
 class Install extends Migration
 {
-    // Public Properties
-    // =========================================================================
-
-    /**
-     * @var string The database driver to use
-     */
-    public $driver;
-
-    // Public Methods
-    // =========================================================================
-
     /**
      * @inheritdoc
      */
     public function safeUp()
     {
-        $this->driver = Craft::$app->getConfig()->getDb()->driver;
-        if ($this->createTables()) {
-            $this->createIndexes();
-            $this->addForeignKeys();
-            // Refresh the db schema caches
-            Craft::$app->db->schema->refresh();
-        }
+        $this->createTables();
+        $this->createIndexes();
+        $this->addForeignKeys();
 
         return true;
     }
@@ -52,48 +31,38 @@ class Install extends Migration
      */
     protected function createTables()
     {
-        $tablesCreated = false;
-
-        $tableSchema = Craft::$app->db->schema->getTableSchema('{{%oauthclient_tokens}}');
-        if ($tableSchema === null) {
-            $this->createTable(
-                '{{%oauthclient_tokens}}',
-                [
-                    'id' => $this->primaryKey(),
-                    'dateCreated' => $this->dateTime()->notNull(),
-                    'dateUpdated' => $this->dateTime()->notNull(),
-                    'uid' => $this->uid(),
-                    'userId' => $this->integer()->notNull(),
-                    'appId' => $this->integer()->notNull(),
-                    'accessToken' => $this->text()->notNull(),
-                    'refreshToken' => $this->text(),
-                    'expiryDate' => $this->dateTime(),
-                ]
-            );
-        }
-
-        $tableSchema = Craft::$app->db->schema->getTableSchema('{{%oauthclient_apps}}');
-        if ($tableSchema === null) {
-            $this->createTable(
-                '{{%oauthclient_apps}}',
-                [
-                    'id' => $this->primaryKey(),
-                    'dateCreated' => $this->dateTime()->notNull(),
-                    'dateUpdated' => $this->dateTime()->notNull(),
-                    'uid' => $this->uid(),
-                    'userId' => $this->integer(),
-                    'name' => $this->string(255)->notNull(),
-                    'handle' => $this->string(255)->notNull(),
-                    'provider' => $this->string(255)->notNull(),
-                    'clientId' => $this->text(),
-                    'clientSecret' => $this->text(),
-                    'urlAuthorize' => $this->text(),
-                    'scopes' => $this->text(),
-                ]
-            );
-        }
-
-        return true;
+        $this->createTable(
+            '{{%oauthclient_tokens}}',
+            [
+                'id' => $this->primaryKey(),
+                'userId' => $this->integer(),
+                'appId' => $this->integer()->notNull(),
+                'accessToken' => $this->text()->notNull(),
+                'refreshToken' => $this->text(),
+                'values' => $this->text(),
+                'expiryDate' => $this->dateTime(),
+                'dateCreated' => $this->dateTime()->notNull(),
+                'dateUpdated' => $this->dateTime()->notNull(),
+                'uid' => $this->uid(),
+            ]
+        );
+        $this->createTable(
+            '{{%oauthclient_apps}}',
+            [
+                'id' => $this->primaryKey(),
+                'dateCreated' => $this->dateTime()->notNull(),
+                'dateUpdated' => $this->dateTime()->notNull(),
+                'uid' => $this->uid(),
+                'userId' => $this->integer(),
+                'name' => $this->string(255)->notNull(),
+                'handle' => $this->string(255)->notNull(),
+                'provider' => $this->string(255)->notNull(),
+                'clientId' => $this->text(),
+                'clientSecret' => $this->text(),
+                'urlAuthorize' => $this->text(),
+                'scopes' => $this->text(),
+            ]
+        );
     }
 
     // Protected Methods
@@ -105,22 +74,14 @@ class Install extends Migration
     protected function createIndexes()
     {
         $this->createIndex(
-            $this->db->getIndexName(
-                '{{%oauthclient_apps}}',
-                'handle',
-                true
-            ),
+            $this->db->getIndexName('{{%oauthclient_apps}}', 'handle', true),
             '{{%oauthclient_apps}}',
             'handle',
             true
         );
 
         $this->createIndex(
-            $this->db->getIndexName(
-                '{{%oauthclient_apps}}',
-                'provider',
-                false
-            ),
+            $this->db->getIndexName('{{%oauthclient_apps}}', 'provider', false),
             '{{%oauthclient_apps}}',
             'provider',
             false
@@ -133,9 +94,7 @@ class Install extends Migration
     protected function addForeignKeys()
     {
         $this->addForeignKey(
-            $this->db->getForeignKeyName('{{%oauthclient_tokens}}', 'appId'),
-            '{{%oauthclient_tokens}}',
-            'appId',
+            $this->db->getForeignKeyName('{{%oauthclient_tokens}}', 'appId'), '{{%oauthclient_tokens}}', 'appId',
             '{{%oauthclient_apps}}',
             'id',
             'CASCADE',
@@ -143,8 +102,7 @@ class Install extends Migration
         );
 
         $this->addForeignKey(
-            $this->db->getForeignKeyName('{{%oauthclient_tokens}}', 'userId'),
-            '{{%oauthclient_tokens}}',
+            $this->db->getForeignKeyName('{{%oauthclient_tokens}}', 'userId'), '{{%oauthclient_tokens}}',
             'userId',
             '{{%users}}',
             'id',
@@ -153,8 +111,7 @@ class Install extends Migration
         );
 
         $this->addForeignKey(
-            $this->db->getForeignKeyName('{{%oauthclient_apps}}', 'userId'),
-            '{{%oauthclient_apps}}',
+            $this->db->getForeignKeyName('{{%oauthclient_apps}}', 'userId'), '{{%oauthclient_apps}}',
             'userId',
             '{{%users}}',
             'id',
@@ -168,15 +125,11 @@ class Install extends Migration
      */
     public function safeDown()
     {
-        $this->driver = Craft::$app->getConfig()->getDb()->driver;
         $this->removeTables();
 
         return true;
     }
 
-    /**
-     * @return void
-     */
     protected function removeTables()
     {
         $this->dropTableIfExists('{{%oauthclient_tokens}}');